bookList refactor
