+++ src/app/components/BookList/components/Book/styles.js
